Added Data Error Preparser
Added a function that returns a list of errors in submitted data.
The CLI now preparses user data for errors and writes them to standard error. This can be suppressed with the --no_inspect (-N) flag.
Added unit tests for the error preparser along with dirty datasets.
Fixed a bug with simulator caused when no cases are generated.